minimal lockwatcher version of motion

+++ video_freebsd.h
@@ -1,10 +1,9 @@
-/*
- * video_freebsd.h
+/*    video_freebsd.h
  *
  *    Include file for video_freebsd.c
- *    Copyright 2004 by Angel Carpintero (motiondevelop@gmail.com)
- *    This software is distributed under the GNU public license version 2
- *    See also the file 'COPYING'.
+ *      Copyright 2004 by Angel Carpintero (ack@telefonica.net)
+ *      This software is distributed under the GNU public license version 2
+ *      See also the file 'COPYING'.
  *
  */
 
@@ -25,13 +24,13 @@
 #include <dev/bktr/ioctl_bt848.h>
 #endif
 
-#endif /* !WITHOUT_V4L */
+#endif
 
 /* bktr (video4linux) stuff FIXME more modes not only these */
 
 /* not used yet FIXME ! only needed for tuner use */
 /*
-#define TV_INPUT_NTSCM    BT848_IFORM_F_NTSCM
+#define TV_INPUT_NTSCM    BT848_IFORM_F_NTSCM 
 #define TV_INPUT_NTSCJ    BT848_IFORM_F_NTSCJ
 #define TV_INPUT_PALBDGHI BT848_IFORM_F_PALBDGHI
 #define TV_INPUT_PALM     BT848_IFORM_F_PALM
@@ -41,26 +40,26 @@
 */
 
 /* video4linux error codes */
-#define V4L_GENERAL_ERROR  0x01   /* binary 000001 */
-#define V4L_BTTVLOST_ERROR 0x05   /* binary 000101 */
+#define V4L_GENERAL_ERROR    0x01  /* binary 000001 */
+#define V4L_BTTVLOST_ERROR   0x05  /* binary 000101 */
 #define V4L_FATAL_ERROR      -1
 
-#define NORM_DEFAULT    0x00800 // METEOR_FMT_AUTOMODE
-#define NORM_PAL        0x00200 // METEOR_FMT_PAL
-#define NORM_NTSC       0x00100 // METEOR_FMT_NTSC
-#define NORM_SECAM      0x00400 // METEOR_FMT_SECAM
-#define NORM_PAL_NC     0x00200 // METEOR_FMT_PAL /* Greyscale howto ?! FIXME */
+#define NORM_DEFAULT      0x00800  /* METEOR_FMT_AUTOMODE */
+#define NORM_PAL          0x00200  /* METEOR_FMT_PAL      */
+#define NORM_NTSC         0x00100  /* METEOR_FMT_NTSC     */
+#define NORM_SECAM        0x00400  /* METEOR_FMT_SECAM    */
+#define NORM_PAL_NC       0x00200  /* METEOR_FMT_PAL      */
 
-#define NORM_DEFAULT_NEW      BT848_IFORM_F_AUTO
-#define NORM_PAL_NEW          BT848_IFORM_F_PALBDGHI
+#define NORM_DEFAULT_NEW      BT848_IFORM_F_AUTO 
+#define NORM_PAL_NEW          BT848_IFORM_F_PALBDGHI    
 #define NORM_NTSC_NEW         BT848_IFORM_F_NTSCM
-#define NORM_SECAM_NEW        BT848_IFORM_F_SECAM
-#define NORM_PAL_NC_NEW       BT848_IFORM_F_AUTO /* FIXME */
+#define NORM_SECAM_NEW        BT848_IFORM_F_SECAM                 
+#define NORM_PAL_NC_NEW       BT848_IFORM_F_AUTO /* FIXME */    
 
 #define PAL                   0
 #define NTSC                  1
 #define SECAM                 2
-#define PAL_NC                3
+#define PAL_NC                3    
 
 #define PAL_HEIGHT          576
 #define SECAM_HEIGHT        576
@@ -77,7 +76,7 @@
 #define BSD_VIDFMT_LAST       8
 
 
-#define IN_DEFAULT            0
+#define IN_DEFAULT            0 
 #define IN_COMPOSITE          0
 #define IN_TV                 1
 #define IN_COMPOSITE2         2
@@ -96,8 +95,8 @@ struct video_dev {
     int fd_tuner;
     const char *video_device;
     const char *tuner_device;
-    unsigned input;
-    unsigned norm;
+    unsigned short input;
+    unsigned short norm;
     int width;
     int height;
     int contrast;
@@ -112,10 +111,10 @@ struct video_dev {
     pthread_mutexattr_t attr;
     int owner;
     int frames;
-
+    
     /* Device type specific stuff: */
-#ifndef WITHOUT_V4L
-     int capture_method;
+#ifndef WITHOUT_V4L    
+     int capture_method;    
     int v4l_fmt;
     unsigned char *v4l_buffers[2];
     int v4l_curbuffer;
